Refactoring + Better notification content format
