+++ Communication/INotificationSender.cs
@@ -1,9 +1,11 @@
 using System.Threading.Tasks;
 
+using SteamGiveawaysBot.Server.Service.Models;
+
 namespace SteamGiveawaysBot.Server.Communication
 {
     public interface INotificationSender
     {
-        Task SendNotification(string content);
+        Task SendNotificationAsync(Reward reward);
     }
 }
